requirements.txt
gradio==4.36.0
python-dotenv==1.0.1
spark-ai-python
pymongo
